Partial builder refactor for better error handling
Adds a ContainerBuildError exception to be thrown on builder failures.
The Task Manager catches this and sets task states to BUILD_FAIL.
