+++ beeflow/common/crt/singularity_driver.py
@@ -4,7 +4,7 @@ Creates text for tasks using Singularity.
 """
 
 from beeflow.common.crt.crt_driver import (ContainerRuntimeDriver, ContainerRuntimeResult, Command)
-from beeflow.common.build.build_driver import task2arg
+from beeflow.common.build.utils import task2arg
 
 
 class SingularityDriver(ContainerRuntimeDriver):
